fix(platform): always mount org-manage dialog in sidebar layout
Move OrgManageDialog from ClerkOrgSwitcher (only rendered in non-collapsed
sidebar) to SidebarLayoutInner so it is always mounted regardless of sidebar
state. On mobile the sidebar starts collapsed and only renders the narrow icon
bar, which excluded ClerkOrgSwitcher entirely — so clicking Invite set the
open signal but no dialog was mounted to respond.

Also open the dialog immediately before awaiting initProfileName$ so the modal
appears instantly on click.
